refactor: move Object interface in root directory
